consumer/requirements.txt
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.2.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.10.0
tzdata==2024.1
